pyproject.toml
[project]
name = "github-issue-previewer"
version = "0.1.4"
description = "A lightweight tool to preview GitHub Issue templates from YAML in real-time with Jinja2 and Watchdog."
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "jinja2>=3.1.6",
    "pyyaml>=6.0.3",
    "typer>=0.20.0",
    "watchdog>=6.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=src --cov-report=term-missing"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]

[project.scripts]
github-issue-previewer = "src.github_issue_previewer.cli:main"

[tool.hatch.build.targets.wheel]
include = [
    "src/github_issue_previewer/**",
    "src/github_issue_previewer/style_template/*.css",
    "src/github_issue_previewer/style_template/*.html"
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true
